requirements.txt
feedparser==6.0.12
requests==2.32.5
pyyaml==6.0.3
python-dotenv==1.2.1
schedule==1.2.2
